+++ playback/src/convert.rs
@@ -8,8 +8,12 @@ pub struct i24([u8; 3]);
 impl i24 {
     fn from_s24(sample: i32) -> Self {
         // trim the padding in the most significant byte
-        let [a, b, c, _d] = sample.to_le_bytes();
-        i24([a, b, c])
+        #[allow(unused_variables)]
+        let [a, b, c, d] = sample.to_ne_bytes();
+        #[cfg(target_endian = "little")]
+        return Self([a, b, c]);
+        #[cfg(target_endian = "big")]
+        return Self([b, c, d]);
     }
 }
 
@@ -30,20 +34,24 @@ impl Converter {
         }
     }
 
-    // Denormalize and dither
-    pub fn scale(&mut self, sample: f32, factor: i64) -> f32 {
+    const SCALE_S32: f64 = 2147483648.;
+    const SCALE_S24: f64 = 8388608.;
+    const SCALE_S16: f64 = 32768.;
+
+    pub fn scale(&mut self, sample: f64, factor: f64) -> f64 {
+        let dither = match self.ditherer {
+            Some(ref mut d) => d.noise(),
+            None => 0.0,
+        };
+
         // From the many float to int conversion methods available, match what
         // the reference Vorbis implementation uses: sample * 32768 (for 16 bit)
-        let int_value = sample * factor as f32;
+        let int_value = sample * factor + dither;
 
-        // https://doc.rust-lang.org/nomicon/casts.html: casting float to integer
-        // rounds towards zero, then saturates. Ideally halves should round to even to
-        // prevent any bias, but since it is extremely unlikely that a float has
-        // *exactly* .5 as fraction, this should be more than precise enough.
-        match self.ditherer {
-            Some(ref mut d) => int_value + d.noise(int_value),
-            None => int_value,
-        }
+        // Casting float to integer rounds towards zero by default, i.e. it
+        // truncates, and that generates larger error than rounding to nearest.
+        // Absolute lowest error is gained from rounding ties to even.
+        math::round::half_to_even(int_value, 0)
     }
 
     // Special case for samples packed in a word of greater bit depth (e.g.
@@ -51,12 +59,12 @@ impl Converter {
     // byte is zero. Otherwise, dithering may cause an overflow. This is not
     // necessary for other formats, because casting to integer will saturate
     // to the bounds of the primitive.
-    pub fn clamping_scale(&mut self, sample: f32, factor: i64) -> f32 {
+    pub fn clamping_scale(&mut self, sample: f64, factor: f64) -> f64 {
         let int_value = self.scale(sample, factor);
 
         // In two's complement, there are more negative than positive values.
-        let min = -factor as f32;
-        let max = (factor - 1) as f32;
+        let min = -factor;
+        let max = factor - 1.0;
 
         if int_value < min {
             return min;
@@ -66,38 +74,42 @@ impl Converter {
         int_value
     }
 
-    pub fn f32_to_s32(&mut self, samples: &[f32]) -> Vec<i32> {
+    pub fn f64_to_f32(&mut self, samples: &[f64]) -> Vec<f32> {
+        samples.iter().map(|sample| *sample as f32).collect()
+    }
+
+    pub fn f64_to_s32(&mut self, samples: &[f64]) -> Vec<i32> {
         samples
             .iter()
-            .map(|sample| self.scale(*sample, 0x80000000) as i32)
+            .map(|sample| self.scale(*sample, Self::SCALE_S32) as i32)
             .collect()
     }
 
     // S24 is 24-bit PCM packed in an upper 32-bit word
-    pub fn f32_to_s24(&mut self, samples: &[f32]) -> Vec<i32> {
+    pub fn f64_to_s24(&mut self, samples: &[f64]) -> Vec<i32> {
         samples
             .iter()
-            .map(|sample| self.clamping_scale(*sample, 0x800000) as i32)
+            .map(|sample| self.clamping_scale(*sample, Self::SCALE_S24) as i32)
             .collect()
     }
 
     // S24_3 is 24-bit PCM in a 3-byte array
-    pub fn f32_to_s24_3(&mut self, samples: &[f32]) -> Vec<i24> {
+    pub fn f64_to_s24_3(&mut self, samples: &[f64]) -> Vec<i24> {
         samples
             .iter()
             .map(|sample| {
                 // Not as DRY as calling f32_to_s24 first, but this saves iterating
                 // over all samples twice.
-                let int_value = self.clamping_scale(*sample, 0x800000) as i32;
+                let int_value = self.clamping_scale(*sample, Self::SCALE_S24) as i32;
                 i24::from_s24(int_value)
             })
             .collect()
     }
 
-    pub fn f32_to_s16(&mut self, samples: &[f32]) -> Vec<i16> {
+    pub fn f64_to_s16(&mut self, samples: &[f64]) -> Vec<i16> {
         samples
             .iter()
-            .map(|sample| self.scale(*sample, 0x8000) as i16)
+            .map(|sample| self.scale(*sample, Self::SCALE_S16) as i16)
             .collect()
     }
 }