+++ playback/src/dither.rs
@@ -32,7 +32,7 @@ pub trait Ditherer {
     where
         Self: Sized;
     fn name(&self) -> &'static str;
-    fn noise(&mut self, sample: f32) -> f32;
+    fn noise(&mut self) -> f64;
 }
 
 impl fmt::Display for dyn Ditherer {
@@ -48,7 +48,7 @@ impl fmt::Display for dyn Ditherer {
 
 pub struct TriangularDitherer {
     cached_rng: ThreadRng,
-    distribution: Triangular<f32>,
+    distribution: Triangular<f64>,
 }
 
 impl Ditherer for TriangularDitherer {
@@ -64,7 +64,7 @@ impl Ditherer for TriangularDitherer {
         Self::NAME
     }
 
-    fn noise(&mut self, _sample: f32) -> f32 {
+    fn noise(&mut self) -> f64 {
         self.distribution.sample(&mut self.cached_rng)
     }
 }
@@ -75,7 +75,7 @@ impl TriangularDitherer {
 
 pub struct GaussianDitherer {
     cached_rng: ThreadRng,
-    distribution: Normal<f32>,
+    distribution: Normal<f64>,
 }
 
 impl Ditherer for GaussianDitherer {
@@ -91,7 +91,7 @@ impl Ditherer for GaussianDitherer {
         Self::NAME
     }
 
-    fn noise(&mut self, _sample: f32) -> f32 {
+    fn noise(&mut self) -> f64 {
         self.distribution.sample(&mut self.cached_rng)
     }
 }
@@ -102,9 +102,9 @@ impl GaussianDitherer {
 
 pub struct HighPassDitherer {
     active_channel: usize,
-    previous_noises: [f32; NUM_CHANNELS],
+    previous_noises: [f64; NUM_CHANNELS],
     cached_rng: ThreadRng,
-    distribution: Uniform<f32>,
+    distribution: Uniform<f64>,
 }
 
 impl Ditherer for HighPassDitherer {
@@ -121,7 +121,7 @@ impl Ditherer for HighPassDitherer {
         Self::NAME
     }
 
-    fn noise(&mut self, _sample: f32) -> f32 {
+    fn noise(&mut self) -> f64 {
         let new_noise = self.distribution.sample(&mut self.cached_rng);
         let high_passed_noise = new_noise - self.previous_noises[self.active_channel];
         self.previous_noises[self.active_channel] = new_noise;